[Refactor/core] 회원 상태변경 이력을 admin.member에서 domain.member로 이동
- MemberStatusHistory, MemberStatusAction, MemberStatusHistoryRepository를 domain.member 패키지로 이동했습니다. 테이블/컬럼 변경은 없습니다.
- domain 계층이 admin 계층을 의존하는 구조를 피하기 위함입니다. 회원 자진 탈퇴 기능에서 이 감사로그를 기록해야 하는데, 그 시점에 domain 쪽 코드가 admin 쪽 클래스를 참조하게 되는 걸 막았습니다.

IssueNum #328
