+++ devine-core/src/main/java/com/umc/devine/domain/member/enums/MemberStatusAction.java
@@ -1,4 +1,4 @@
-package com.umc.devine.admin.member.enums;
+package com.umc.devine.domain.member.enums;
 
 public enum MemberStatusAction {
     SUSPEND, UNSUSPEND, FORCE_WITHDRAW, CANCEL_WITHDRAWAL, WITHDRAWAL_FINALIZED
