+++ devine-core/src/test/java/com/umc/devine/domain/member/repository/MemberStatusHistoryRepositoryTest.java
@@ -1,11 +1,10 @@
-package com.umc.devine.admin.member.repository;
+package com.umc.devine.domain.member.repository;
 
-import com.umc.devine.admin.member.entity.MemberStatusHistory;
-import com.umc.devine.admin.member.enums.MemberStatusAction;
+import com.umc.devine.domain.member.entity.MemberStatusHistory;
+import com.umc.devine.domain.member.enums.MemberStatusAction;
 import com.umc.devine.domain.member.entity.Member;
 import com.umc.devine.domain.member.enums.MemberMainType;
 import com.umc.devine.domain.member.enums.MemberStatus;
-import com.umc.devine.domain.member.repository.MemberRepository;
 import com.umc.devine.support.CoreIntegrationTestSupport;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
